GPKG: fixed mapping of DbColNumber to INTEGER

+++ src/ch/ehi/sqlgen/generator_impl/jdbc/GeneratorGeoPackage.java
@@ -79,7 +79,7 @@ public class GeneratorGeoPackage extends GeneratorJdbc {
 			type="TEXT(36)"; // 550e8400-e29b-11d4-a716-446655440000
 		}else if(column instanceof DbColNumber){
 			DbColNumber col=(DbColNumber)column;
-			type="DOUBLE";
+			type="INTEGER";
 		}else if(column instanceof DbColVarchar){
 			int colsize=((DbColVarchar)column).getSize();
 			if(colsize==-1){
